+++ policyfiles/default.rb
@@ -3,5 +3,3 @@ name 'default'
 include_policy 'base', path: './base.lock.json'
 
 run_list 'managed-chef-server::default'
-
-override['mcs']['skip_test'] = true