+++ policyfiles/legacy.rb
@@ -8,4 +8,4 @@ override['mcs']['cookbooks']['dir'] = '/backups/cookbooks'
 override['mcs']['environments']['dir'] = '/backups/environments'
 override['mcs']['roles']['dir'] = '/backups/roles'
 
-default['mcs']['skip_test'] = true
+override['mcs']['skip_test'] = true