+++ test/integration/legacy/default_test.rb
@@ -4,7 +4,7 @@
 
 # check output of knife commands
 
-describe command('knife environment list -c /etc/opscode/managed/config.rb') do
+describe command('knife environment list -c /etc/opscode/managed/chef_managed_org/config.rb') do
   its ('stdout') { should match /^_default$/ }
   its ('stdout') { should match /^essex$/ }
   its ('stdout') { should match /^lab$/ }
@@ -12,29 +12,28 @@ describe command('knife environment list -c /etc/opscode/managed/config.rb') do
   its ('stdout') { should_not match /^lab-admin$/ }
 end
 
-describe command('knife role list -c /etc/opscode/managed/config.rb') do
+describe command('knife role list -c /etc/opscode/managed/chef_managed_org/config.rb') do
   its ('stdout') { should match /^base$/ }
   its ('stdout') { should match /^lab-admin$/ }
   its ('stdout') { should match /^lab-base$/ }
   its ('stdout') { should_not match /^lab-environment$/ }
 end
 
-describe command('knife cookbook list -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /^apt           7.1.1$/ }
-  its ('stdout') { should match /^chef-client   11.0.3$/ }
-  its ('stdout') { should match /^cron          6.2.1$/ }
-  its ('stdout') { should match /^iptables      4.3.4$/ }
-  its ('stdout') { should match /^logrotate     2.2.0$/ }
-  its ('stdout') { should match /^mattray       0.8.0$/ }
-  its ('stdout') { should match /^ntp           3.6.0$/ }
-  its ('stdout') { should match /^openssh       2.7.0$/ }
-  its ('stdout') { should match /^sudo          5.5.0$/ }
+describe command('knife cookbook list -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /^apt\s+7.1.1$/ }
+  its ('stdout') { should match /^chef-client\s+11.0.3$/ }
+  its ('stdout') { should match /^cron\s+6.2.1$/ }
+  its ('stdout') { should match /^iptables\s+4.3.4$/ }
+  its ('stdout') { should match /^logrotate\s+2.2.0$/ }
+  its ('stdout') { should match /^mattray\s+0.8.0$/ }
+  its ('stdout') { should match /^ntp\s+3.6.0/ }
+  its ('stdout') { should match /^sudo\s+5.5.0/ }
 end
 
-describe command('knife cookbook show ntp -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /^ntp   3.6.0  3.5.0  3.4.0$/ }
+describe command('knife cookbook show ntp -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /^ntp.*3.6.0/ }
 end
 
-describe command('knife cookbook show sudo -c /etc/opscode/managed/config.rb') do
-  its ('stdout') { should match /^sudo   5.5.0  5.4.0$/ }
+describe command('knife cookbook show sudo -c /etc/opscode/managed/chef_managed_org/config.rb') do
+  its ('stdout') { should match /^sudo.*5.4.0$/ }
 end